requirements.txt
# Core GCP dependencies
google-cloud-aiplatform>=1.133.0
google-cloud-storage>=3.8.0
google-auth>=2.47.0
openai>=2.15.0
requests>=2.32.5

# FastAPI web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# Development tools
jupyterlab>=4.5.2
ipykernel>=7.1.0
matplotlib>=3.10.8
tensorflow>=2.20.0
